[B] Restart Puma without SIGUSR2
While it would be nice to restart Puma with SIGUSR2, this introduces a
problem during upgrades. In all cases, if we send a restart signal to
Puma during a reconfigure run, we want Puma to start whether or not it
is currently running. Sending a SIGUSR2 signal will only restart Puma if
it is currently running. In the case of database migrations, the puma
services will have stopped, and are unable to respond to SIGUSR2 when we
send the restart signal. In short, this commit helps ensure that the
API processes start up correctly after migrations run.